HZN-973: Fix failing tests
